requirements.txt
google-generativeai
requests
Pillow
google-auth-oauthlib
google-api-python-client
